apify/twshaw3/storelocator/edeka_de/requirements.txt
lxml==4.3.4
requests==2.11.1
# uncomment if you use selenium
# selenium==3.141.0
sglogging
sgzip==0.0.55